+++ src/utils/mod.rs
@@ -202,7 +202,7 @@ pub use csv::{OhlcvCandle, read_ohlcv_from_zip};
 pub use logger::{setup_logger, setup_logger_with_level};
 pub use others::{
     DETERMINISTIC_RNG_DEFAULT_SEED, approx_equal, deterministic_rng, get_random_element,
-    process_n_times_iter, random_decimal,
+    random_decimal,
 };
 pub use time::TimeFrame;
 pub use traits::Len;